Changed data type for wall location to float

+++ include/gas_container.h
@@ -21,8 +21,8 @@ class GasContainer {
    */
   GasContainer(const int window_size, const int num_of_particles);
 
-  GasContainer(const int left_wall_loc, const int top_wall_loc,
-               const int right_wall_loc, const int bottom_wall_loc, const int num_of_particles);
+  GasContainer(const float left_wall_loc, const float top_wall_loc,
+               const float right_wall_loc, const float bottom_wall_loc, const int num_of_particles);
 
   /**
    * Displays the container walls and the current positions of the particles_.
